+++ R/citation.R
@@ -2,33 +2,29 @@
 #'
 #' This function uses \code{\link[utils]{citation}} and
 #' \code{\link[utils]{toBibtex}} to create bib entries for R packages and write
-#' them in a file. Only the auto-generated citations are included for a package.
-#' This function can facilitate the auto-generation of bibliography databases
-#' for R packages, and it is easy to regenerate all the citations after updating
-#' R packages.
+#' them in a file. It can facilitate the auto-generation of bibliography
+#' databases for R packages, and it is easy to regenerate all the citations
+#' after updating R packages.
 #'
-#' The citation is forced to be generated from the \file{DESCRIPTION} file of
-#' the package (\code{citation(auto = TRUE)}). The keyword \samp{R-pkgname} is
-#' used for the bib item, where \samp{pkgname} is the name of the package. All
-#' references specified in the \file{CITATION} file of the package are ignored.
-#' The main purpose of this function is to automate the generation of the
-#' package citation information because it often changes (e.g. author, year,
-#' package version, ...). By comparison, paper/book citations don't change too
-#' often, so it is not a lot of work even if you just cut and paste such
-#' bibliography entries from \code{toBibtex(citation())}.
+#' For a package, the keyword \samp{R-pkgname} is used for its bib item, where
+#' \samp{pkgname} is the name of the package. Citation entries specified in the
+#' \file{CITATION} file of the package are also included. The main purpose of
+#' this function is to automate the generation of the package citation
+#' information because it often changes (e.g. author, year, package version,
+#' ...).
 #'
 #' @param x Package names. Packages which are not installed are ignored.
 #' @param file The (\file{.bib}) file to write. By default, or if \code{NULL},
 #'   output is written to the R console.
 #' @param tweak Whether to fix some known problems in the citations, especially
 #'   non-standard format of author names.
-#' @param width Width of lines in bibliography entries. If \code{NULL},
-#'   lines will not be wrapped.
+#' @param width Width of lines in bibliography entries. If \code{NULL}, lines
+#'   will not be wrapped.
 #' @param prefix Prefix string for keys in BibTeX entries; by default, it is
 #'   \samp{R-} unless \code{\link{option}('knitr.bib.prefix')} has been set to
 #'   another string.
-#' @return A list containing the citations. Citations are also written to the \code{file} as a
-#'   side effect.
+#' @return A list containing the citations. Citations are also written to the
+#'   \code{file} as a side effect.
 #' @note Some packages on CRAN do not have standard bib entries, which was once
 #'   reported by Michael Friendly at
 #'   \url{https://stat.ethz.ch/pipermail/r-devel/2010-November/058977.html}. I
@@ -69,6 +65,7 @@ write_bib = function(
     x = x[!idx]
   }
   x = setdiff(x, .base.pkgs) # remove base packages
+  x = sort(x)
   bib = sapply(x, function(pkg) {
     cite = citation(pkg, auto = if (pkg == 'base') NULL else TRUE)
     if (tweak) {
@@ -98,7 +95,27 @@ write_bib = function(
       structure(c(b[idx[1L]], b[-idx], b[idx[2L]]), class = 'Bibtex')
     })
   }
-  bib = bib[sort(x)]
+  # also read citation entries from the CITATION file if provided
+  bib2 = lapply(x, function(pkg) {
+    if (pkg == 'base') return()
+    if (system.file('CITATION', package = pkg) == '') return()
+    cites = citation(pkg, auto = FALSE)
+    cites = Filter(x = cites, function(cite) {
+      # exclude entries identical to citation(pkg, auto = TRUE)
+      !isTRUE(grepl('R package version', cite$note))
+    })
+    s = make_unique(unlist(lapply(cites, function(cite) {
+      if (is.null(cite$year)) format(Sys.Date(), '%Y') else cite$year
+    })))
+    mapply(cites, s, FUN = function(cite, suffix) {
+      # the entry is likely to be the same as citation(pkg, auto = TRUE)
+      if (isTRUE(grepl('R package version', cite$note))) return()
+      entry = toBibtex(cite)
+      entry[1] = sub('\\{,$', sprintf('{%s%s,', pkg, suffix), entry[1])
+      entry
+    }, SIMPLIFY = FALSE)
+  })
+  bib = c(bib, unlist(bib2, recursive = FALSE))
   if (!is.null(file) && length(x)) write_utf8(unlist(bib), file)
   invisible(bib)
 }